+++ source/src/Tile.cpp
@@ -1,7 +1,7 @@
 #include <SDL2/SDL.h>
 
-#include "Graphics.h"
-#include "Tile.h"
+#include "../headers/Graphics.h"
+#include "../headers/Tile.h"
 
 Tile::Tile() {}
 
@@ -13,8 +13,9 @@ Tile::Tile(SDL_Texture* tileset, Vector2 size, Vector2 tilesetPosition, Vector2 
 void Tile::Update(float elapsedTime) {}
 
 void Tile::Draw(Graphics &graphics) {
-	SDL_Rect destRect = { this->_mapPosition.x, this->_mapPosition.y, this->_size.x * globals::SPRITE_SCALE,
-						this->_size.y * globals::SPRITE_SCALE };
+	SDL_Rect destRect = { this->_mapPosition.x, this->_mapPosition.y,
+						  static_cast<int>(this->_size.x * globals::SPRITE_SCALE),
+						static_cast<int>(this->_size.y * globals::SPRITE_SCALE) };
 	SDL_Rect sourceRect = { this->_tilesetPosition.x, this->_tilesetPosition.y,
 			this->_size.x, this->_size.y };
 	graphics.DrawSurface(this->_tileset, &sourceRect, &destRect);
